requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-jose[cryptography]==3.5.0
httpx==0.28.1
requests==2.32.4
mcp==1.12.1